move odds fetching to nba_data_service
