+++ backend/src/nba_wins_pool/services/nba_data_service.py
@@ -318,6 +318,47 @@ class NbaDataService:
         )
         return game_df
 
+    def get_fanduel_moneyline_odds(self) -> dict[str, dict[str, float]]:
+        """Fetch vig-adjusted FanDuel moneyline win probabilities for today's games.
+
+        Returns:
+            Dict mapping game_id -> {"home": float, "away": float} win probabilities,
+            or empty dict if the request fails.
+        """
+        url = "https://cdn.nba.com/static/json/liveData/odds/odds_todaysGames.json"
+        try:
+            response = requests.get(url, timeout=10)
+            response.raise_for_status()
+            data = response.json()
+        except Exception:
+            logger.warning("Failed to fetch FanDuel odds from %s", url, exc_info=True)
+            return {}
+
+        result = {}
+        for game in data.get("games", []):
+            game_id = str(game.get("gameId", ""))
+            moneyline = next((m for m in game.get("markets", []) if m.get("name") == "2way"), None)
+            if not moneyline:
+                continue
+            fanduel = next((b for b in moneyline.get("books", []) if b.get("name") == "FanDuel"), None)
+            if not fanduel:
+                continue
+            outcomes = fanduel.get("outcomes", [])
+            home_out = next((o for o in outcomes if o.get("type") == "home"), None)
+            away_out = next((o for o in outcomes if o.get("type") == "away"), None)
+            if not home_out or not away_out:
+                continue
+            home_odds = home_out.get("odds")
+            away_odds = away_out.get("odds")
+            if not home_odds or not away_odds:
+                continue
+            raw_home = 1 / float(home_odds)
+            raw_away = 1 / float(away_odds)
+            total = raw_home + raw_away
+            result[game_id] = {"home": raw_home / total, "away": raw_away / total}
+
+        return result
+
     async def _store_data(self, key: str, data: dict) -> None:
         """Store data in database cache (generic helper).
