+++ backend/tests/test_leaderboard_service.py
@@ -1,7 +1,4 @@
-import json
 from datetime import date
-from pathlib import Path
-from unittest.mock import MagicMock, patch
 from uuid import uuid4
 
 import pandas as pd
@@ -11,8 +8,6 @@ from nba_wins_pool.services.leaderboard_service import LeaderboardService
 from nba_wins_pool.services.nba_data_service import NBAGameStatus
 from nba_wins_pool.services.pool_season_service import TeamRosterMappings
 from nba_wins_pool.types.season_str import SeasonStr
-
-FIXTURES_DIR = Path(__file__).parent / "fixtures"
 
 UNDRAFTED = "Undrafted"
 
@@ -331,7 +326,7 @@ def _make_today_games_service(game_df, teams_data):
         auction_valuation_service=None,
     )
     # Avoid live HTTP calls in tests that don't exercise odds logic
-    service._fetch_fanduel_moneyline_odds = lambda: {}
+    service.nba_data_service.get_fanduel_moneyline_odds = lambda: {}
     return service
 
 
@@ -636,136 +631,6 @@ async def test_today_games_empty_when_no_games():
     assert result == []
 
 
-# ---------------------------------------------------------------------------
-# _fetch_fanduel_moneyline_odds tests
-# ---------------------------------------------------------------------------
-
-
-def _make_bare_service() -> LeaderboardService:
-    """Minimal LeaderboardService instance for testing odds parsing."""
-    return LeaderboardService(
-        db_session=None,
-        pool_repository=None,
-        roster_repository=None,
-        roster_slot_repository=None,
-        team_repository=None,
-        nba_data_service=None,
-        pool_season_service=None,
-        auction_valuation_service=None,
-    )
-
-
-def _mock_requests_get(fixture_data: dict):
-    mock_response = MagicMock()
-    mock_response.json.return_value = fixture_data
-    mock_response.raise_for_status.return_value = None
-    return mock_response
-
-
-def test_fetch_fanduel_odds_parses_fixture():
-    """Vig-adjusted FanDuel probabilities are computed correctly from the fixture."""
-    fixture = json.loads((FIXTURES_DIR / "sample-nba-odds-today-response.json").read_text())
-    service = _make_bare_service()
-
-    with patch("nba_wins_pool.services.leaderboard_service.requests.get", return_value=_mock_requests_get(fixture)):
-        result = service._fetch_fanduel_moneyline_odds()
-
-    # Game 0022501042: FanDuel home 4.200, away 1.247
-    assert "0022501042" in result
-    odds = result["0022501042"]
-    raw_home = 1 / 4.200
-    raw_away = 1 / 1.247
-    total = raw_home + raw_away
-    assert odds["home"] == pytest.approx(raw_home / total, rel=1e-6)
-    assert odds["away"] == pytest.approx(raw_away / total, rel=1e-6)
-    assert odds["home"] + odds["away"] == pytest.approx(1.0, rel=1e-6)
-
-
-def test_fetch_fanduel_odds_all_probabilities_sum_to_one():
-    """Every game returned sums to 1.0."""
-    fixture = json.loads((FIXTURES_DIR / "sample-nba-odds-today-response.json").read_text())
-    service = _make_bare_service()
-
-    with patch("nba_wins_pool.services.leaderboard_service.requests.get", return_value=_mock_requests_get(fixture)):
-        result = service._fetch_fanduel_moneyline_odds()
-
-    assert len(result) > 0
-    for game_id, odds in result.items():
-        assert odds["home"] + odds["away"] == pytest.approx(1.0, rel=1e-6), f"game {game_id} does not sum to 1"
-
-
-def test_fetch_fanduel_odds_skips_game_without_fanduel():
-    """Games with no FanDuel book are excluded from results."""
-    data = {
-        "games": [
-            {
-                "gameId": "0022501099",
-                "markets": [
-                    {
-                        "name": "2way",
-                        "books": [
-                            {
-                                "name": "Novibet",
-                                "outcomes": [
-                                    {"type": "home", "odds": "2.000"},
-                                    {"type": "away", "odds": "2.000"},
-                                ],
-                            }
-                        ],
-                    }
-                ],
-            }
-        ]
-    }
-    service = _make_bare_service()
-
-    with patch("nba_wins_pool.services.leaderboard_service.requests.get", return_value=_mock_requests_get(data)):
-        result = service._fetch_fanduel_moneyline_odds()
-
-    assert "0022501099" not in result
-
-
-def test_fetch_fanduel_odds_skips_game_without_2way_market():
-    """Games with only a spread market and no 2way market are excluded."""
-    data = {
-        "games": [
-            {
-                "gameId": "0022501099",
-                "markets": [
-                    {
-                        "name": "spread",
-                        "books": [
-                            {
-                                "name": "FanDuel",
-                                "outcomes": [
-                                    {"type": "home", "odds": "1.909", "spread": "-5.5"},
-                                    {"type": "away", "odds": "1.909", "spread": "5.5"},
-                                ],
-                            }
-                        ],
-                    }
-                ],
-            }
-        ]
-    }
-    service = _make_bare_service()
-
-    with patch("nba_wins_pool.services.leaderboard_service.requests.get", return_value=_mock_requests_get(data)):
-        result = service._fetch_fanduel_moneyline_odds()
-
-    assert "0022501099" not in result
-
-
-def test_fetch_fanduel_odds_returns_empty_on_http_error():
-    """Network failures return an empty dict without raising."""
-    service = _make_bare_service()
-
-    with patch("nba_wins_pool.services.leaderboard_service.requests.get", side_effect=Exception("connection refused")):
-        result = service._fetch_fanduel_moneyline_odds()
-
-    assert result == {}
-
-
 @pytest.mark.asyncio
 async def test_today_games_includes_odds_for_pregame():
     """Pregame games include vig-adjusted win percentages from FanDuel."""
@@ -811,7 +676,7 @@ async def test_today_games_includes_odds_for_pregame():
     service = _make_today_games_service(game_df, teams_data)
 
     fixed_odds = {"0022501042": {"home": 0.2289, "away": 0.7711}}
-    service._fetch_fanduel_moneyline_odds = lambda: fixed_odds
+    service.nba_data_service.get_fanduel_moneyline_odds = lambda: fixed_odds
 
     result = await service.get_today_games(uuid4(), SeasonStr("2025-26"))
 
@@ -864,7 +729,7 @@ async def test_today_games_odds_null_when_fanduel_unavailable():
         },
     ]
     service = _make_today_games_service(game_df, teams_data)
-    service._fetch_fanduel_moneyline_odds = lambda: {}
+    service.nba_data_service.get_fanduel_moneyline_odds = lambda: {}
 
     result = await service.get_today_games(uuid4(), SeasonStr("2025-26"))
